refactor(frontend): split Settings.jsx into per-section modules
Settings.jsx (802 lines) becomes a ~65-line orchestrator; each section
(Oura, AI provider, MCP clients, cloud sync, data import/export) moves
to frontend/src/screens/settings/ with shared input styles. No behavior
changes — mechanical extraction with imports adjusted.
